refactor: rename sudt balance to udt balance
